requirements.txt
torch>=1.13.0
torchvision>=0.14.0
timm>=0.6.12
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tqdm>=4.62.0
grad-cam>=1.4.0